requirements.txt
pip~=24.3.1
redis~=5.2.1
typing_extensions~=4.12.2
rich~=13.9.4
Pygments~=2.19.1
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.1.8
Jinja2~=3.1.5
blinker~=1.9.0
itsdangerous~=2.2.0
mdurl~=0.1.2
requests~=2.32.3
Deprecated~=1.2.15
limits~=4.0.0
packaging~=24.2
certifi~=2024.12.14
urllib3~=2.3.0
idna~=3.10
wrapt~=1.17.1
Flask~=3.1.0
python-dotenv~=1.0.1
Flask-Login~=0.6.3
flask-httpauth~=4.8.0
